Custom site addons
Created a new PHP file in `includes` for any custom site code and a
style sheet in `styles/custom/css` for styles added for a specific
site. This way, if you keep copies of these in a safe place, you can
more safely update the theme. Also fixed quote bug for captions, and
changed writing form so it only displays featured image and intro text
for first view of the form (so you can add really long instructions).