Sort Options and Shortcodes
Taglist shortcode provides a way to list all tags used, set the sorting, and optionally limit number of tags shown

TRU Options provides settings to change the sorting of published items to be by date or alphabetical, and can be ascending or descending. There are also settings to apply it only to the home page, or just to tag and/or category archives. Also updated template to properly display archive headings and descriptions (for tag and category archives)